Clarify an ambiguity in initialization of INT_ZERO_EXP
git-svn-id: https://svn.code.sf.net/p/plexil/code/branches/plexil-4@4111 d2be67b0-814c-0410-949f-d215b6b29573

+++ src/expr/ExpressionConstants.cc
@@ -55,7 +55,7 @@ namespace PLEXIL
 
   Expression *INT_ZERO_EXP()
   {
-    static IntegerConstant sl_int_zero(0);
+    static IntegerConstant sl_int_zero((int32_t) 0);
     return &sl_int_zero;
   }
